+++ WindowsFormsApp2/Form1.cs
@@ -27,7 +27,7 @@ namespace WindowsFormsApp2
         }
         public UserControl UC2
         {
-            get { return userControl21; }
+            get { return userControl22; }
         }
         public UserControl UC3
         {
@@ -35,7 +35,7 @@ namespace WindowsFormsApp2
         }
         public RichTextBox rtb
         {
-            get { return userControl21.richTextBox1; }
+            get { return userControl22.richTextBox1; }
         }
         public int selectLength
         {
@@ -54,7 +54,7 @@ namespace WindowsFormsApp2
         }
 
         //userControl11 --> home
-        //userControl21 --> the analyzerS
+        //userControl22 --> the analyzerS
         private void Home_Click(object sender, EventArgs e)
         {
             // Changing the BackColor of "Home" Button
@@ -64,7 +64,7 @@ namespace WindowsFormsApp2
             // Changing the BackColor of "Settings" Button
             Settings.BackColor = System.Drawing.ColorTranslator.FromHtml("#ffffff");
             userControl11.Show();
-            userControl21.Hide();
+            userControl22.Hide();
             userControl31.Hide();
             userControl11.BringToFront();
         }
@@ -74,10 +74,10 @@ namespace WindowsFormsApp2
             Analyze.BackColor = System.Drawing.ColorTranslator.FromHtml("#DDDDDD");
             Home.BackColor = System.Drawing.ColorTranslator.FromHtml("#ffffff");
             Settings.BackColor = System.Drawing.ColorTranslator.FromHtml("#ffffff");
-            userControl21.Show();
+            userControl22.Show();
             userControl11.Hide();
             userControl31.Hide();
-            userControl21.BringToFront();
+            userControl22.BringToFront();
         }
 
         private void Settings_Click(object sender, EventArgs e)
@@ -87,7 +87,7 @@ namespace WindowsFormsApp2
             Analyze.BackColor = System.Drawing.ColorTranslator.FromHtml("#ffffff");
             userControl31.Show();
             userControl11.Hide();
-            userControl21.Hide();
+            userControl22.Hide();
             userControl31.BringToFront();
         }
 
@@ -96,10 +96,10 @@ namespace WindowsFormsApp2
             //Properties.Appearance.Default.Reset();
             //Properties.Appearance.Default.Save();
 
-            this.Text = "Artificial Ignorance (AI) Text Analyzer " + version;
-            versionlabel.Text = "AI " + version + "\r\nBy Hugo Lau\r\n";
             this.Home_Click(sender, e);
             this.Form1_Resize(sender, e);
+            this.Text = "Artificial Ignorance (AI) Text Analyzer " + version;
+            versionlabel.Text = "AI " + version + "\r\nBy Hugo Lau\r\n";
             //Analyze.Font = new Font("UD Digi Kyokasho NK-B", 24, FontStyle.Bold);
             Home.Font = new Font(Properties.Appearance.Default.FontFamily, 15, FontStyle.Bold);
             Analyze.Font = new Font(Properties.Appearance.Default.FontFamily, 15, FontStyle.Bold);
@@ -109,7 +109,7 @@ namespace WindowsFormsApp2
         }
 
 
-        private void userControl21_Load(object sender, EventArgs e)
+        private void userControl22_Load(object sender, EventArgs e)
         {
         }
 
@@ -127,22 +127,18 @@ namespace WindowsFormsApp2
             //panel1.Size = new Size(formWidth/9, panel1.Height);
 
             userControl11.Location = new Point(200, 0);
-            userControl21.Location = new Point(200, 0);
+            userControl22.Location = new Point(200, 0);
             userControl31.Location = new Point(200, 0);
 
             userControl11.Size = new Size(this.Width - 200, panel1.Height);
-            userControl21.Size = new Size(this.Width - 200, panel1.Height);
+            userControl22.Size = new Size(this.Width - 200, panel1.Height);
             userControl31.Size = new Size(this.Width - 200, panel1.Height);
 
             Home.Location = new Point(panel1.Size.Width / 2 - Home.Size.Width / 2, 100);
             Analyze.Location = new Point(panel1.Size.Width / 2 - Home.Size.Width / 2, Home.Location.Y + Home.Size.Height + 70);
             Settings.Location = new Point(panel1.Size.Width / 2 - Home.Size.Width / 2, Analyze.Location.Y + Analyze.Size.Height + 70);
+            close.Location = new Point(panel1.Size.Width / 2 - Home.Size.Width / 2, Settings.Location.Y + Settings.Size.Height + 70);
             versionlabel.Location = new System.Drawing.Point(10, this.Size.Height - 80);
-        }
-
-        private void homeToolTip_Popup(object sender, PopupEventArgs e)
-        {
-
         }
 
         private void userControl31_Load(object sender, EventArgs e)
@@ -159,13 +155,18 @@ namespace WindowsFormsApp2
         {
             var color = Properties.Appearance.Default.BackColor;
             userControl11.BackColor = System.Drawing.ColorTranslator.FromHtml($"#{color}");
-            userControl21.BackColor = System.Drawing.ColorTranslator.FromHtml($"#{color}");
+            userControl22.BackColor = System.Drawing.ColorTranslator.FromHtml($"#{color}");
             userControl31.BackColor = System.Drawing.ColorTranslator.FromHtml($"#{color}");
 
             var colorx = System.Drawing.ColorTranslator.FromHtml($"#{color}");
             var newcolor = Color.FromArgb(Math.Min(255, Int32.Parse(colorx.R.ToString()) + 62), Math.Min(Int32.Parse(colorx.G.ToString()) + 48, 255), Math.Min(Int32.Parse(colorx.B.ToString()) + 33, 255));
             panel1.BackColor = newcolor;
             versionlabel.BackColor = newcolor;
+        }
+
+        private void close_Click(object sender, EventArgs e)
+        {
+            this.Close();
         }
 
         public void changeFont()
@@ -176,9 +177,5 @@ namespace WindowsFormsApp2
             versionlabel.Font = new Font(Properties.Appearance.Default.FontFamily, 10, FontStyle.Bold);
         }
 
-        private void versionTextBox_TextChanged(object sender, EventArgs e)
-        {
-
-        }
     }
 }