+++ WindowsFormsApp2/changelog.md
@@ -5,10 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.0] - 2023-09-08
+## [0.8.0] - 2023-09-10
 
 ### Added
 - Search & Replace button & functionality
+- Added readability table in PDF report
+- Added mispelled word count into popup form & Pdf report
+- Added close button to main form
+
+### Changed 
+- Changed PDF report format
+
+### Fixed
+- Mispelled red lines now shown properly in PDF report
 
 ## [0.7.0] - 2023-09-05